changed things

+++ planner/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
-# from django.contrib import messages
+from django.contrib import messages
 import calendar
 from calendar import HTMLCalendar
 from datetime import datetime
@@ -12,22 +12,36 @@ from datetime import datetime, timedelta
 
 
 @login_required
-def planner(request):
-    now = datetime.now()
-    year = now.year
-    month_number = now.month
+def planner(request, year=None, month=None):
+    if year is None or month is None:
+        now = datetime.now()
+        year = now.year
+        month = now.month
+    else:
+        month_number = list(calendar.month_name).index(month.title())
+        year = int(year)
+        month = month_number
 
-    # Get events for the current month
-    events = Event.objects.filter(event_date__year=year, event_date__month=month_number)
+    # Calculate previous and next month
+    first_day = datetime(year, month, 1)
+    last_month = (first_day - timedelta(days=1)).replace(day=1)
+    next_month = (first_day + timedelta(days=32)).replace(day=1)
+    
+    # Get events for the specified month
+    events = Event.objects.filter(event_date__year=year, event_date__month=month)
 
     # Create calendar
-    cal = EventCalendar(events).formatmonth(year, month_number)
+    cal = EventCalendar(events).formatmonth(year, month)
 
     return render(request, 'planner.html', {
         "name": request.user,
         "year": year,
-        "month": now.strftime('%B'),
-        "cal": cal
+        "month": calendar.month_name[month],
+        "cal": cal,
+        "prev_year": last_month.year,
+        "prev_month": last_month.strftime('%B'),
+        "next_year": next_month.year,
+        "next_month": next_month.strftime('%B')
     })
 
 @login_required
@@ -101,5 +115,10 @@ def edit_event(request, event_id):
 @login_required
 def delete_event(request, event_id):
     event = get_object_or_404(Event, pk=event_id)
-    event.delete()
+    if request.user == event.manager:
+        event.delete()
+        messages.success(request, ("Event Deleted"))
+    else:
+        messages.success(request, ("You can't do that"))
+    
     return redirect('planner:all_events')
